fix: #352 - Fixed reading relative files from tsconfig.json with multiple rootDirs
Previously, when relative file paths were passed as "files" in tsconfig,
and there were multiple root directories, all combinations of rootdir
and file path was created. This was incorrect, as the purpose of having
multiple root dirs is to have your files in one of these dirs.

To fix that, TsConfigResolver now checks if the file/subdirectory exists
before adding it to the list of files/subdirectories.