feat: #456 - Relax ts dependency to 3.0.1 and 3.0.3.
The typescript declaration file is no longer stored internally in the library.

I would like to widen support for 3.1.x, but that compiler version seems slightly broken. It resolves a lot of types to `any` within this library. I will need to investigate and report the issue to the typescript team if it hasn't been reported (or perhaps it's an issue internally in the library).