performance: #702 - Fix Project#addSourceFileAtPaths being needlessly slow in some scenarios.
BREAKING CHANGE: Due to the fix for #702, when using `Project#addSourceFileAtPaths` directories that do not have an ancestor directory with a source file included in the results will no longer be added to the project. If you want to ensure that a directory and all its subfolders are added, use `Project#addDirectoryAtPath(path, { recursive: true })`.