fix - Handle comma separated definitions with DefinitionInfo#getDeclarationNode
The current implementation has an off-by-one bug that causes getDeclarationNode()
to return undefined instead of the defition when the definition is immediately
preceeded by a "," (comma token).

For example, "myFunction" in: `const someConst=23,myFunction=function () {};`

This change adjusts the lower bound of the AST travesal to not include the end
position, which in the case of comma separated variable declarations with no
whitespace results in the recursion incorrectly identifying the comma token as
the declaration node.